microservices/portal-ms/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
django-environ
psycopg2-binary
dj-database-url
Pillow          
django-storages
pytest
pytest-django
requests-mock
ruff 